requirements.txt
celery==5.4.0
Django==4.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.1
django-modeltranslation==0.19.10
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gevent==24.10.3
pillow==11.0.0
psycopg2-binary==2.9.10
redis==5.1.1
setuptools==75.2.0
hitcount==1.3.5